+++ compiler/light-classes/src/org/jetbrains/kotlin/asJava/KtLightAnnotation.kt
@@ -18,6 +18,7 @@ package org.jetbrains.kotlin.asJava
 
 import com.intellij.openapi.util.TextRange
 import com.intellij.psi.*
+import com.intellij.psi.impl.source.resolve.reference.ReferenceProvidersRegistry
 import com.intellij.util.IncorrectOperationException
 import org.jetbrains.kotlin.idea.KotlinLanguage
 import org.jetbrains.kotlin.psi.KtAnnotationEntry
@@ -66,6 +67,8 @@ class KtLightAnnotation(
             }
         }
 
+        override fun getReference() = references.singleOrNull()
+        override fun getReferences() = ReferenceProvidersRegistry.getReferencesFromProviders(delegate, PsiReferenceService.Hints.NO_HINTS)
         override fun getLanguage() = KotlinLanguage.INSTANCE
     }