+++ idea/testData/spring/core/gutter/classGutterAbstractShowMappedInheritors/AbstractBeanInheritor.kt
@@ -1,0 +1,1 @@
+class AbstractBeanInheritor : AbstractBean()